third_party/blink/web_tests/svg/dynamic-updates/SVGCircleElement-dom-cx-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'cx' attribute of the SVGCircleElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS circleElement.getAttribute('cx') is "-50"
PASS circleElement.getAttribute('cx') is "150"
PASS successfullyParsed is true

TEST COMPLETE
